resolvido problema na requisição de buscar um produto e funcionalidades prévias da api foram finalizadas

+++ src/main/java/com/api/produtos/controller/ProdutoController.java
@@ -2,10 +2,14 @@ package com.api.produtos.controller;
 
 import javax.validation.Valid;
 
+import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
@@ -15,7 +19,7 @@ import com.api.produtos.model.ProdutoModel;
 import com.api.produtos.repository.ProdutoRepository;
 
 @RestController
-@RequestMapping("/produto")
+@RequestMapping("/api/produto")
 public class ProdutoController {
 	
 	@Autowired
@@ -33,10 +37,41 @@ public class ProdutoController {
 		return produtoRepository.save(produto);
 	}
 	
-	@DeleteMapping
-	public ProdutoModel deletarProduto(@RequestBody ProdutoModel produto) {
+	@DeleteMapping("/{id}")
+	public ResponseEntity<Void> excluir(@PathVariable Long id){
+		ProdutoModel produto = produtoRepository.getOne(id);
+		
+		if(produto == null) {
+			return ResponseEntity.notFound().build();
+		}
+		
 		produtoRepository.delete(produto);
 		
-		return produto;
+		return ResponseEntity.noContent().build();
+	}
+	
+	@GetMapping("/{id}")
+	public ResponseEntity<ProdutoModel> buscar(@PathVariable Long id){
+		ProdutoModel produto = produtoRepository.getOne(id);
+		
+		if(produto == null) {
+			return ResponseEntity.notFound().build();
+		}
+		
+		return ResponseEntity.ok(produto);
+	}
+	
+	@PutMapping("/{id}")
+	public ResponseEntity<ProdutoModel> atualizar(@PathVariable Long id, @Valid @RequestBody ProdutoModel produto){
+		ProdutoModel produtoExistente = produtoRepository.getOne(id);
+		
+		if(produtoExistente == null) {
+			return ResponseEntity.notFound().build();
+		}
+		
+		BeanUtils.copyProperties(produto, produtoExistente, "id");
+		produtoExistente = produtoRepository.save(produtoExistente);
+		
+		return ResponseEntity.ok(produtoExistente);
 	}
 }
